+++ src/main/java/com/multicolorshulkers/client/ModConfig.java
@@ -4,8 +4,6 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.multicolorshulkers.MultiColorShulkers;
 import net.fabricmc.loader.api.FabricLoader;
-import net.minecraft.client.util.InputUtil;
-import org.lwjgl.glfw.GLFW;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -16,9 +14,9 @@ public class ModConfig {
 	private static final Path CONFIG_PATH = FabricLoader.getInstance().getConfigDir().resolve("dual-dye-shulkers.json");
 
 	private static ModConfig INSTANCE;
-
-	// Config options
 	public boolean showTooltip = true;
+	public boolean enableCrafting = true;
+	public boolean enableKeybinds = true;
 
 	// Top binding: Shift + Right Click
 	public String topKey1 = "key.keyboard.left.shift";
@@ -28,65 +26,11 @@ public class ModConfig {
 	public String bottomKey1 = "key.keyboard.left.control";
 	public String bottomKey2 = "key.mouse.right";
 
-	// Transient cached combos
-	private transient KeyCombo cachedTopCombo;
-	private transient KeyCombo cachedBottomCombo;
-
-	// Default combos
-	public static final KeyCombo DEFAULT_TOP_COMBO = new KeyCombo(
-			InputUtil.Type.KEYSYM.createFromCode(GLFW.GLFW_KEY_LEFT_SHIFT),
-			InputUtil.Type.MOUSE.createFromCode(GLFW.GLFW_MOUSE_BUTTON_RIGHT)
-	);
-	public static final KeyCombo DEFAULT_BOTTOM_COMBO = new KeyCombo(
-			InputUtil.Type.KEYSYM.createFromCode(GLFW.GLFW_KEY_LEFT_CONTROL),
-			InputUtil.Type.MOUSE.createFromCode(GLFW.GLFW_MOUSE_BUTTON_RIGHT)
-	);
-
 	public static ModConfig get() {
 		if (INSTANCE == null) {
 			INSTANCE = load();
 		}
 		return INSTANCE;
-	}
-
-	public KeyCombo getTopCombo() {
-		if (cachedTopCombo == null) {
-			try {
-				InputUtil.Key k1 = InputUtil.fromTranslationKey(topKey1);
-				InputUtil.Key k2 = InputUtil.fromTranslationKey(topKey2);
-				cachedTopCombo = new KeyCombo(k1, k2);
-			} catch (Exception e) {
-				MultiColorShulkers.LOGGER.warn("Invalid top combo, using default");
-				cachedTopCombo = DEFAULT_TOP_COMBO;
-			}
-		}
-		return cachedTopCombo;
-	}
-
-	public void setTopCombo(KeyCombo combo) {
-		this.cachedTopCombo = combo;
-		this.topKey1 = combo.getKey1().getTranslationKey();
-		this.topKey2 = combo.getKey2().getTranslationKey();
-	}
-
-	public KeyCombo getBottomCombo() {
-		if (cachedBottomCombo == null) {
-			try {
-				InputUtil.Key k1 = InputUtil.fromTranslationKey(bottomKey1);
-				InputUtil.Key k2 = InputUtil.fromTranslationKey(bottomKey2);
-				cachedBottomCombo = new KeyCombo(k1, k2);
-			} catch (Exception e) {
-				MultiColorShulkers.LOGGER.warn("Invalid bottom combo, using default");
-				cachedBottomCombo = DEFAULT_BOTTOM_COMBO;
-			}
-		}
-		return cachedBottomCombo;
-	}
-
-	public void setBottomCombo(KeyCombo combo) {
-		this.cachedBottomCombo = combo;
-		this.bottomKey1 = combo.getKey1().getTranslationKey();
-		this.bottomKey2 = combo.getKey2().getTranslationKey();
 	}
 
 	public static ModConfig load() {